[lcn] Fix displaying of "not enough licenses" message (#9761)
