+++ bundles/org.openhab.binding.lcn/src/main/java/org/openhab/binding/lcn/internal/connection/ConnectionStateShutdown.java
@@ -13,6 +13,7 @@
 package org.openhab.binding.lcn.internal.connection;
 
 import org.eclipse.jdt.annotation.NonNullByDefault;
+import org.eclipse.jdt.annotation.Nullable;
 import org.openhab.binding.lcn.internal.common.LcnAddr;
 
 /**
@@ -40,6 +41,11 @@ public class ConnectionStateShutdown extends AbstractConnectionState {
 
     @Override
     public void onPckMessageReceived(String data) {
+        // nothing
+    }
+
+    @Override
+    public void handleConnectionFailed(@Nullable Throwable e) {
         // nothing
     }
 }
